### 2020-02-01 #2052 @whedon
[PRE REVIEW]: The pdb2sql Python Package: Parsing, Manipulation and Analysis of PDB Files Using SQL Queries
**Submitting author:** @NicoRenaud (<a href="http://orcid.org/0000-0001-9589-2694">Nicolas Renaud</a>)
**Repository:** <a href="https://github.com/DeepRank/pdb2sql" target ="_blank">https://github.com/DeepRank/pdb2sql</a>
**Version:** 0.2.2
