### 2020-02-03 b6f0e38
Create CITATION.cff
citation metadata
