+++ CITATION.cff
@@ -1,0 +1,26 @@
+# YAML 1.2
+---
+abstract: "Fast and versatile PDB parser using SQL queries."
+authors: 
+  -
+    affiliation: "Netherlands eScience Center"
+    family-names: Renaud
+    given-names: Nicolas
+    orcid: "https://orcid.org/0000-0001-9589-2694"
+  -
+    affiliation: "Netherlands eScience Center"
+    family-names: Geng
+    given-names: Cunliang
+    orcid: "https://orcid.org/0000-0002-1409-8358"
+cff-version: "1.1.0"
+doi: "10.5281/zenodo.3232887"
+keywords: 
+  - pdb
+  - sql
+  - "protein structure"
+  - biomolecule
+license: "Apache-2.0"
+message: "If you use this software, please cite it using these metadata."
+repository-code: "https://github.com/DeepRank/pdb2sql"
+title: pdb2sql
+...
