### 2020-02-03 #2052 @NicoRenaud
Dear editor, I’m not sure if I should answer that email as it is my first submission to JOSS. But I’ve fixed the missing DOI and all the DOIs should be in order now. ThanksBest wishesNico  From: Kevin Mattheus MoermanSent: Saturday, 1 February 2020 23:15To: openjournals/joss-reviewsCc: Nicolas Renaud; MentionSubject: Re: [openjournals/joss-reviews] [PRE REVIEW]: The pdb2sql Python Package: Parsing, Manipulation and Analysis of PDB Files Using SQL Queries (#2052) @NicoRenaud can you check the potentially missing DOI ☝️, thanks—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
