### 2020-02-04 3f959ec
update citation and version number
