+++ CITATION.cff
@@ -1,7 +1,7 @@
 # YAML 1.2
 ---
-abstract: "Fast and versatile PDB parser using SQL queries."
-authors: 
+abstract: "Fast and versatile biomolecular structure PDB file parser using SQL queries."
+authors:
   -
     affiliation: "Netherlands eScience Center"
     family-names: Renaud
@@ -14,11 +14,12 @@ authors:
     orcid: "https://orcid.org/0000-0002-1409-8358"
 cff-version: "1.1.0"
 doi: "10.5281/zenodo.3232887"
-keywords: 
-  - pdb
-  - sql
+keywords:
+  - "pdb parser"
   - "protein structure"
   - biomolecule
+  - bioinformatics
+  - CAPRI
 license: "Apache-2.0"
 message: "If you use this software, please cite it using these metadata."
 repository-code: "https://github.com/DeepRank/pdb2sql"
