### 2020-02-06 #2052 @NicoRenaud
Dear Lorena, Joao Rodriguez (https://github.com/JoaoRodrigues) would be a good reviewer. He has written pdb_tools which is similar to pdb2sql Best,Nico  From: Lorena PantanoSent: Wednesday, 5 February 2020 23:36To: openjournals/joss-reviewsCc: Nicolas Renaud; MentionSubject: Re: [openjournals/joss-reviews] [PRE REVIEW]: The pdb2sql Python Package: Parsing, Manipulation and Analysis of PDB Files Using SQL Queries (#2052) @NicoRenaud, do you recommend any review for this? I will start the review process and add it afterwards. Thanks!—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or unsubscribe.
