### 2020-02-24 DeepRank/pdb2sql#21 @JoaoRodrigues
Installation: unused dependencies in setup.py
Installing pdb2sql from `pip` downloads a number of third-party modules but not all seem necessary. For example, cython and matplotlib are never used anywhere on the codebase and add quite a bit of space/time to the installation process.

Also, the dependencies listed on the `requirements.txt` file are not synced with those on setup.py. I know these serve different purposes but it would be better if they matched. If you want to have a 'development' requirements file, then I'd name it requirements-dev.txt and add extra dependencies there.

Linking to the review at JOSS: https://github.com/openjournals/joss-reviews/issues/2077
