### 2020-02-24 DeepRank/pdb2sql#22 @JoaoRodrigues
Unclear warning messages on PDB read.
When loading a PDB file (e.g. `1AK4_5w.pdb` from the tests folder) there is a cryptic warning message about missing elements:

```
C:\Users\joaor\Miniconda3\envs\pdb2sql\lib\site-packages\pdb2sql\pdb2sqlcore.py:230: UserWarning: Missing element and guess it with atom type
  warnings.warn("Missing element and guess it with atom type")
```

I would improve the message itself. I have no idea which atoms are missing element columns or if the program guessed the elements (correctly or not).

Link to JOSS review: https://github.com/openjournals/joss-reviews/issues/2077
