### 2020-02-24 DeepRank/pdb2sql#23 @JoaoRodrigues
Inconsistency in imports between documentation and quick example
The `README.md` file suggests `from pdb2sql.pdb2sqlcore import pdb2sql` while the 10-min intro on the docs page suggests (the much simpler!!) `from pdb2sql import pdb2sql`.

Link to JOSS review: https://github.com/openjournals/joss-reviews/issues/2077
