### 2020-02-24 DeepRank/pdb2sql#24 @JoaoRodrigues
CONTRIBUTING.rst is not for this project
It seems you copied the `CONTRIBUTING.rst` file from another repository but didn't edit the links nor the text, e.g.
> push your feature branch to (your fork of) the pyCHAMP repository on GitHub;

Link to JOSS review: https://github.com/openjournals/joss-reviews/issues/2077
