### 2020-02-24 #2077 @JoaoRodrigues
Hey @NicoRenaud

The tool looks good but there are a few rough edges I'd polish before publishing it. I left a few issues on your repo that should be easy to fix. Here are a few more minor questions/issues I found when going over the paper and the code:

- [ ] There is a consistent typo on the word 'structure' all over the documentation :) `sed -e 's/strcture/structure/g'` should fix it!
- [ ] As a user, I would expect I could use the interface module on a PDB I already parsed, e.g. `pdb = pdb2sql('myfile.pdb'); pdbint = interface(pdb)`. Right now I need to load it twice and this is a bit counter-intuitive. Maybe this is a simple fix?
- [ ] I'd keep to one convention when naming variables and attributes. You use a lot of `get_X` methods, for example, but then there are attributes named `nModels`, `chainID`, `altLoc`. You should stick to one, e.g., `n_models`, `chain_id`, `altloc` for simplicity.
- [ ] For convenience, you could add a fetch method to download files from the RCSB PDB directly. Makes it easier for new users to try the tool without having to find a PDB of their own first.
- [ ] The paper states that pdb2sql tries to abstract SQL language for the end-user, and it generally does a pretty good job! However, having methods like `db.close()` and such in the documentation (and exposed in the public API) sort of defeats the purpose. I would make these private and, my preference here, would make them happen behind the scenes so that the end user never has to worry about db connections, cursors, etc.
- [ ] On the topic of dependencies, I see that sqlalchemy is not necessary for the tool to run, but it provides a better OO interface. Maybe making it a runtime dependency then? Or explain the difference between the two interfaces and why you'd want (or not) to use sqlalchemy.

Overall, the tool does what it says it does and it's pretty quick too. As I said before, just some rough edges that need polishing! Good job @NicoRenaud and @CunliangGeng !
