### 2020-02-25 c4d30c3
Delete .setup_save.py
