+++ .setup_save.py
@@ -1,10 +1,0 @@
-from setuptools import setup
-
-setup(
-    name='pdb2sql',
-    description='Using SQL queries to manipulate PDB files',
-    version='0.1-dev0',
-    url=' ',
-    packages=['pdb2sql'],
-    install_requires=['numpy']
-)
