+++ MANIFEST.in
@@ -1,2 +1,4 @@
 include LICENSE
-include README.rst
+include README.md
+graft test/pdb
+include test/utils.py
