+++ pdb2sql/pdb2sql_base.py
@@ -1,6 +1,4 @@
 import os
-from time import time
-
 
 class pdb2sql_base(object):
 
