+++ .travis.yml
@@ -13,14 +13,6 @@ before_install:
   - hash -r
   - conda config --set always_yes yes --set changeps1 no --set auto_update_conda False
 
-  # dependencies
-  - conda install cython numpy
-  - sudo apt-get install sqlite3
-
-
-  # codacy coverage
-  - pip install -q coverage codacy-coverage coveralls
-
 # command to install dependencies
 install:
   - pip install -e .[dev,test]
@@ -36,4 +28,3 @@ script:
 after_script:
 - python-codacy-coverage -r coverage.xml
 - coveralls
-
