### 2020-02-27 a429fbc
change sql close to private method
