+++ pdb2sql/pdb2sql_base.py
@@ -234,7 +234,7 @@ class pdb2sql_base(object):
 
         return i
 
-    def close(self, rmdb=True):
+    def _close(self, rmdb=True):
 
         if self.sqlfile is None:
             self.conn.close()
