### 2020-02-27 bff26e9
remove sql close from doc
