### 2020-02-28 c76494a
Update interface to allow pdb2sql object as input
