### 2020-02-28 ed1fded
Add utils fetch method to download pdb file
