+++ pdb2sql/__init__.py
@@ -2,5 +2,6 @@ from .pdb2sqlcore import pdb2sql
 from .interface import interface
 from .StructureSimilarity import StructureSimilarity
 from . import transform
+from .utils import fetch
 
 from .__version__ import __version__
