### 2020-02-28 ff31317
Change nModel to private in pdb2sqlcore.py
