### 2020-02-28 fd087f0
remove sql close in tests
