### 2020-02-28 5288b7b
Correct paper.md based on grammarly
