### 2020-03-03 293d026
Add tests for interface initilization using pdb2sql instance as input
