+++ test/test_interface.py
@@ -1,5 +1,6 @@
 import unittest
 from pdb2sql import interface
+from pdb2sql import pdb2sql
 
 
 class Test_1_ContactAtoms(unittest.TestCase):
@@ -267,6 +268,33 @@ class Test_2_ContactResidues(unittest.TestCase):
         self.assertEqual(len(contact_residues['L']), 9)
         self.assertEqual(len(contact_residues['R']), 8)
 
+
+class Test_3_PDB2SQLInstanceInput(unittest.TestCase):
+    """test using pdb2sql instance as input"""
+
+    def setUp(self):
+        self.pdb = 'pdb/3CRO.pdb'
+
+    def test_get_contact_residues_default(self):
+        """"verify get_contact_residues default."""
+        pdb_db = pdb2sql(self.pdb)
+        self.db = interface(pdb_db)
+        contact_residues = self.db.get_contact_residues()
+        self.assertIsInstance(contact_residues, dict)
+        self.assertEqual(len(contact_residues), 2)
+        self.assertEqual(list(contact_residues.keys()), ['A', 'B'])
+        self.assertEqual(len(contact_residues['A']), 20)
+        self.assertEqual(len(contact_residues['B']), 20)
+
+    def test_database_consistency(self):
+        """"verify initilizing interface with updated pdb2sql database"""
+        pdb_db = pdb2sql(self.pdb)
+        pdb_db.update_column('temp', [99]*10)
+        target = pdb_db.get('*')
+
+        self.db = interface(pdb_db)
+        result = self.db.get('*')
+        self.assertEqual(target, result)
 
 if __name__ == '__main__':
     runner = unittest.TextTestRunner(verbosity=2)
