+++ pdb2sql/pdb2sqlAlchemy.py
@@ -1,325 +1,0 @@
-from sqlalchemy import Column, Integer, String, Float
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from .pdb2sql_base import pdb2sql_base
-import re
-
-Base = declarative_base()
-
-
-class ATOM(Base):
-    """SQLAlchemy object for atoms."""
-
-    __tablename__ = 'ATOM'
-    rowID = Column(Integer, primary_key=True)
-    serial = Column(Integer, nullable=False)
-    name = Column(String(6), nullable=False)
-    altLoc = Column(String(1), nullable=False)
-    resName = Column(String(3), nullable=False)
-    chainID = Column(String(1), nullable=False)
-    resSeq = Column(Integer, nullable=False)
-    iCode = Column(String(1), nullable=False)
-    x = Column(Float, nullable=False)
-    y = Column(Float, nullable=False)
-    z = Column(Float, nullable=False)
-    occ = Column(Float, nullable=False)
-    temp = Column(Float, nullable=False)
-    element = Column(String(2), nullable=False)
-    model = Column(Integer, nullable=False)
-
-
-class pdb2sql_alchemy(pdb2sql_base):
-
-    def __init__(
-            self,
-            pdbfile,
-            sqlfile=None,
-            fix_chainID=False,
-            verbose=False):
-        """Use sqlAlchemy to load the database."""
-        super().__init__(pdbfile, sqlfile, fix_chainID, verbose)
-        self._create_sql()
-
-    def _create_sql(self):
-
-        # sqlalchemy init
-        self.conn = create_engine('sqlite:///:memory:')
-        Base.metadata.create_all(self.conn)
-        DBSession = sessionmaker(bind=self.conn)
-        self.session = DBSession()
-
-        # read the pdb file a pure python way
-        # RMK we go through the data twice here. Once to read the ATOM line and once to parse the data ...
-        # we could do better than that. But the most time consuming step seems to be the CREATE TABLE query
-        # with open(pdbfile,'r') as fi:
-        #   data = [line.split('\n')[0] for line in fi if line.startswith('ATOM')]
-
-        fi = open(self.pdbfile, 'r')
-        self.nModel = 0
-        _check_format = True
-
-        for line in fi:
-
-            if line.startswith('ATOM'):
-                line = line.split('\n')[0]
-
-            elif line.startswith('ENDMDL'):
-                self.nModel += 1
-                continue
-
-            else:
-                continue
-
-            if _check_format:
-                # old format chain ID fix
-                del_copy = self.delimiter.copy()
-                if line[del_copy['chainID'][0]] == ' ':
-                    del_copy['chainID'] = [72, 73]
-                if line[del_copy['chainID'][0]] == ' ':
-                    raise ValueError('chainID not found sorry')
-                _check_format = False
-
-            # browse all attribute of each atom
-            at = {}
-            for _, (colname, coltype) in enumerate(self.col.items()):
-
-                # get the piece of data
-                if colname in del_copy.keys():
-                    data = line[del_copy[colname][0]
-                        :del_copy[colname][1]].strip()
-
-                    # convert it if necessary
-                    if coltype == 'INT':
-                        data = int(data)
-                    elif coltype == 'REAL':
-                        data = float(data)
-
-                    # append to dict
-                    at[colname] = data
-
-            # create a new ATOM
-            newat = ATOM(
-                serial=at['serial'],
-                name=at['name'],
-                altLoc=at['altLoc'],
-                resName=at['resName'],
-                chainID=at['chainID'],
-                resSeq=at['resSeq'],
-                iCode=at['iCode'],
-                x=at['x'],
[…]
