### 2020-03-04 dc55e5c
remove sqlalchemy from docs
