### 2020-03-04 38f4910
add clean to docs makefile
