### 2020-03-04 6d210c7
add local check to docs makefile
