### 2020-03-04 7d7dc29
Update interface docstring
