### 2020-03-06 #2077 @NicoRenaud
Dear @JoaoRodrigues, thanks for reviewing our code and for your comments and suggestions. We've addressed them all, merged the changes in the master and updated the paper accordingly. As a quick summary :
- [x] we fixed the consistent typos on the documentation, thanks for spotting that
- [x] You can directly use `pdbint=interface('myfile.pdb')` or if needed first create an sql db and pass it to the interface : `pdb = pdb2sql('myfile.pdb)` pdbint =interface(pdb)`
- [x] The fetch function is a very good idea that is now implemented !
- [x] Indeed exposing the `.close()` method is not ideal. This is now fully done behind the scene as you suggested. (Note that we always use in memory db)
- [x] we realized that we never use sqlalchemy and that we do not want to maintain code that is never used. Therefore we have decided to entirely remove sqlalchemy support form pdb2sql. We kept it in a separate branch in case there is a demand for it. We have updated the documentation and the paper accordingly.
Thanks again for all your comments !
