### 2020-03-20 e9a8e59
extracted superpose routines
