+++ pdb2sql/superpose.py
@@ -1,0 +1,210 @@
+
+import numpy as np
+from .pdb2sqlcore import pdb2sql
+from .transform import rotate
+
+def superpose(pdb1,pdb2,chainID=None,method='svd'):
+    """superpose two complexes
+    
+    Arguments:
+        pdb1 {[type]} -- [description]
+        pdb2 {[type]} -- [description]
+    
+    Keyword Arguments:
+        method {str} -- [description] (default: {'svd'})
+    """
+
+    backbone_atoms = ['CA','C','N','O']
+
+    sql1 = pdb2sql(pdb1)
+    sql2 = pdb2sql(pdb2)
+
+    # xyz of the chains selected
+    chain_xyz1 = np.array(sql1.get("x,y,z",chainID=chainID, name = backbone_atoms))
+    chain_xyz2 = np.array(sql2.get("x,y,z",chainID=chainID, name = backbone_atoms))
+
+    # translation vector
+    tr1 = get_trans_vect(chain_xyz1)
+    tr2 = get_trans_vect(chain_xyz2)
+
+    # rotation matrix
+    chain_xyz1 += tr1
+    chain_xyz2 += tr2
+    center = np.mean(chain_xyz2,0)
+    U = get_rotation_matrix(chain_xyz2, chain_xyz1, method=method)
+
+    # transform the coordinate of second pdb
+    xyz2 = np.array(sql2.get("x,y,z"))
+    xyz2 += tr2 
+    xyz2 = rotate(xyz2,U,center=center)
+    xyz2 -= tr1
+
+    # update the second sql
+    sql2.update('x,y,z',xyz2)
+    fname = pdb2.strip('.pdb')+'_aligned_on_'+pdb1.strip('.pdb')+'.pdb'
+    sql2.exportpdb(fname)
+
+# compute the translation vector to center a set of points
+def get_trans_vect(P):
+    """Get the translationv vector to the origin.
+
+    Args:
+        P (np.array(nx3)): position of the points in the molecule
+
+    Returns:
+        float: minus mean value of the xyz columns
+    """
+    return -np.mean(P, 0)
+
+# main switch for the rotation matrix
+# add new methods here if necessary
+def get_rotation_matrix(P, Q, method='svd'):
+
+    # get the matrix with Kabsh method
+    if method.lower() == 'svd':
+        mat = get_rotation_matrix_Kabsh(P, Q)
+
+    # or with the quaternion method
+    elif method.lower() == 'quaternion':
+        mat = get_rotation_matrix_quaternion(P, Q)
+
+    else:
+        raise ValueError(
+            f'{method} is not a valid method for rmsd alignement. '
+            f'Options are svd or quaternions')
+
+    return mat
+
+# get the rotation matrix via a SVD
+# decomposition of the correlation matrix
+def get_rotation_matrix_Kabsh(P, Q):
+    """Get the rotation matrix to aligh two point clouds.
+
+    The method is based on th Kabsh approach
+    https://cnx.org/contents/HV-RsdwL@23/Molecular-Distance-Measures
+
+    Args:
+        P (np.array): xyz of the first point cloud
+        Q (np.array): xyz of the second point cloud
+
+    Returns:
+        np.array: rotation matrix
+
+    Raises:
+        ValueError: matrix have different sizes
+    """
+    pshape = P.shape
[…]
