+++ pdb2sql/StructureSimilarity.py
@@ -2,6 +2,8 @@ import warnings
 import numpy as np
 from .pdb2sqlcore import pdb2sql
 from .interface import interface
+from .superpose import get_trans_vect, get_rotation_matrix
+
 from . import transform
 import os
 import pickle
@@ -135,8 +137,8 @@ class StructureSimilarity(object):
                 self.ref, resData, return_not_in_zone=True)
 
         # get the translation so that both A chains are centered
-        tr_decoy = self.get_trans_vect(xyz_decoy_long)
-        tr_ref = self.get_trans_vect(xyz_ref_long)
+        tr_decoy = get_trans_vect(xyz_decoy_long)
+        tr_ref = get_trans_vect(xyz_ref_long)
 
         # translate everything for 1
         xyz_decoy_short += tr_decoy
@@ -148,7 +150,7 @@ class StructureSimilarity(object):
 
         # get the ideal rotation matrix
         # to superimpose the A chains
-        U = self.get_rotation_matrix(
+        U = get_rotation_matrix(
             xyz_decoy_long, xyz_ref_long, method=method)
 
         # rotate the entire fragment
@@ -282,8 +284,8 @@ class StructureSimilarity(object):
             xyz_contact_ref = self.get_xyz_zone_backbone(self.ref, resData)
 
         # get the translation so that both A chains are centered
-        tr_decoy = self.get_trans_vect(xyz_contact_decoy)
-        tr_ref = self.get_trans_vect(xyz_contact_ref)
+        tr_decoy = get_trans_vect(xyz_contact_decoy)
+        tr_ref = get_trans_vect(xyz_contact_ref)
 
         # translate everything
         xyz_contact_decoy += tr_decoy
@@ -291,7 +293,7 @@ class StructureSimilarity(object):
 
         # get the ideal rotation matrix
         # to superimpose the A chains
-        U = self.get_rotation_matrix(
+        U = get_rotation_matrix(
             xyz_contact_decoy,
             xyz_contact_ref,
             method=method)
@@ -559,8 +561,8 @@ class StructureSimilarity(object):
             xyz_ref_short = xyz_ref_A
 
         # get the translation so that both A chains are centered
-        tr_decoy = self.get_trans_vect(xyz_decoy_long)
-        tr_ref = self.get_trans_vect(xyz_ref_long)
+        tr_decoy = get_trans_vect(xyz_decoy_long)
+        tr_ref = get_trans_vect(xyz_ref_long)
 
         # translate everything for 1
         xyz_decoy_short += tr_decoy
@@ -572,7 +574,7 @@ class StructureSimilarity(object):
 
         # get the ideal rotation matrix
         # to superimpose the A chains
-        U = self.get_rotation_matrix(
+        U = get_rotation_matrix(
             xyz_decoy_long, xyz_ref_long, method=method)
 
         # rotate the entire fragment
@@ -756,8 +758,8 @@ class StructureSimilarity(object):
                 'Error in i-rmsd: only one chain represented in one chain')
 
         # get the translation so that both A chains are centered
-        tr_decoy = self.get_trans_vect(xyz_contact_decoy)
-        tr_ref = self.get_trans_vect(xyz_contact_ref)
+        tr_decoy = get_trans_vect(xyz_contact_decoy)
+        tr_ref = get_trans_vect(xyz_contact_ref)
 
         # translate everything
         xyz_contact_decoy += tr_decoy
@@ -765,7 +767,7 @@ class StructureSimilarity(object):
 
         # get the ideql rotation matrix
[…]
