### 2020-03-20 98f04cd
created align routines and test
