+++ test/test_align.py
@@ -2,6 +2,7 @@ import unittest
 from pdb2sql.align import align, pca
 import numpy as np
 
+
 class TestAlign(unittest.TestCase):
     """Test the superpose functionality"""
 
@@ -11,7 +12,7 @@ class TestAlign(unittest.TestCase):
     def test_align(self):
         """test the routine."""
 
-        for idir, axis in zip([0,1,2], ['x','y','z']):
+        for idir, axis in zip([0, 1, 2], ['x', 'y', 'z']):
             sql = align(self.pdb, axis=axis)
             xyz = np.array(sql.get('x,y,z'))
             u, v = pca(xyz)
