+++ pdb2sql/interface.py
@@ -2,7 +2,7 @@ import numpy as np
 import itertools
 import warnings
 from .pdb2sqlcore import pdb2sql
-
+from .pdb2sql.align import 
 
 class interface(pdb2sql):
 
@@ -300,3 +300,4 @@ class interface(pdb2sql):
                     set([tuple(resData) for resData in data[chain]]))
 
             return residue_contact
+
