### 2020-03-23 e67f41b
fix bugs for superpose
