+++ test/test_align.py
@@ -1,5 +1,5 @@
 import unittest
-from pdb2sql.align import align, pca
+from pdb2sql.align import align, pca, align_interface
 import numpy as np
 
 
@@ -18,3 +18,22 @@ class TestAlign(unittest.TestCase):
             u, v = pca(xyz)
             vmax = np.abs(v[:, np.argmax(u)])
             assert np.argmax(vmax) == idir
+
+    def test_align_interface(self):
+        """test the routine."""
+
+        def get_xyz_interface(sql):
+            idx = sql.get_contact_atoms()
+            idx = idx['A'] + idx['B']
+            return np.array(sql.get('x,y,z', rowID=idx)) 
+
+        for idir, plane in zip([2, 1, 0], ['xy', 'xz', 'yz']):
+            sql = align_interface(self.pdb, plane=plane)
+            xyz = get_xyz_interface(sql)
+            u, v = pca(xyz)
+            vmax = np.abs(v[:, np.argmin(u)])
+            assert np.argmax(vmax) == idir
+
+    
+
+
