### 2020-03-27 a7552ad
renamed transform_xyz_from ... to superpose_selection
