### 2020-03-27 18e62c6
fix docstring
