+++ test/test_structureSimilarity.py
@@ -46,26 +46,22 @@ class TestSim(unittest.TestCase):
 
     def test_irmsdsql_default(self):
         """verify compute_irmsd_pdb2sql()"""
-        with self.assertWarns(UserWarning) as ex:
-            result = self.sim.compute_irmsd_pdb2sql()
+        result = self.sim.compute_irmsd_pdb2sql()
         self.assertEqual(result, self.irmsd)
 
     def test_irmsdsql_izone(self):
         """verify compute_irmsd_pdb2sql(izone='sql.izone)"""
-        with self.assertWarns(UserWarning) as ex:
-            result = self.sim.compute_irmsd_pdb2sql(izone=self.izone)
+        result = self.sim.compute_irmsd_pdb2sql(izone=self.izone)
         self.assertEqual(result, self.irmsd)
 
     def test_irmssql_method(self):
         """verify compute_irmsd_pdb2sql(method='quaternion')"""
-        with self.assertWarnsRegex(UserWarning, 'Element is missing'):
-            result = self.sim.compute_irmsd_pdb2sql(method='quaternion')
+        result = self.sim.compute_irmsd_pdb2sql(method='quaternion')
         self.assertEqual(result, self.irmsd)
 
     def test_irmsdsql_exportpdb(self):
         """verify compute_irmsd_pdb2sql(exportpath='.')"""
-        with self.assertWarns(UserWarning) as ex:
-            result = self.sim.compute_irmsd_pdb2sql(exportpath='.')
+        result = self.sim.compute_irmsd_pdb2sql(exportpath='.')
         self.assertEqual(result, self.irmsd)
         self.assertTrue(os.path.isfile('./irmsd_ref.pdb'))
         self.assertTrue(os.path.isfile('./irmsd_decoy.pdb'))
@@ -100,20 +96,17 @@ class TestSim(unittest.TestCase):
 
     def test_lrmsdsql_default(self):
         """verify compute_lrmsd_pdb2sql()"""
-        with self.assertWarns(UserWarning) as ex:
-            result = self.sim.compute_lrmsd_pdb2sql()
+        result = self.sim.compute_lrmsd_pdb2sql()
         self.assertEqual(result, self.lrmsd)
 
     def test_lrmsdsql_method(self):
         """verify compute_lrmsd_pdb2sql(method='quaternion')"""
-        with self.assertWarns(UserWarning):
-            result = self.sim.compute_lrmsd_pdb2sql(method='quaternion')
+        result = self.sim.compute_lrmsd_pdb2sql(method='quaternion')
         self.assertEqual(result, self.lrmsd)
 
     def test_lrmsdsql_exportpdb(self):
         """verify compute_lrmsd_pdb2sql(exportpath='.')"""
-        with self.assertWarnsRegex(UserWarning, 'Element is missing'):
-            result = self.sim.compute_lrmsd_pdb2sql(exportpath='.')
+        result = self.sim.compute_lrmsd_pdb2sql(exportpath='.')
         self.assertEqual(result, self.lrmsd)
         self.assertTrue(os.path.isfile('./lrmsd_ref.pdb'))
         self.assertTrue(os.path.isfile('./lrmsd_decoy.pdb'))
@@ -132,8 +125,7 @@ class TestSim(unittest.TestCase):
 
     def test_fnatsql_default(self):
         """verify compute_fnat_pdb2sql()"""
-        with self.assertWarnsRegex(UserWarning, 'Element is missing'):
-            result = self.sim.compute_fnat_pdb2sql()
+        result = self.sim.compute_fnat_pdb2sql()
         self.assertEqual(result, self.fnat)
 
     ####################################################################
