### 2020-03-28 7caefc3
update warnings and errors
