+++ test/test_align.py
@@ -10,27 +10,31 @@ class TestAlign(unittest.TestCase):
         self.pdb = 'pdb/1AK4/1AK4_10w.pdb'
 
     def test_align(self):
-        """test the routine."""
+        """Test align()"""
 
         for idir, axis in zip([0, 1, 2], ['x', 'y', 'z']):
-            sql = align(self.pdb, axis=axis)
-            xyz = np.array(sql.get('x,y,z'))
+            with self.assertWarns(UserWarning) as ex:
+                db = align(self.pdb, axis=axis)
+            xyz = np.array(db.get('x,y,z'))
             u, v = pca(xyz)
             vmax = np.abs(v[:, np.argmax(u)])
             assert np.argmax(vmax) == idir
 
     def test_align_interface(self):
-        """test the routine."""
+        """Test align_interface()"""
 
-        def get_xyz_interface(sql):
-            idx = sql.get_contact_atoms()
+        def get_xyz_interface(db):
+            idx = db.get_contact_atoms()
             idx = idx['A'] + idx['B']
-            return np.array(sql.get('x,y,z', rowID=idx)) 
+            return np.array(db.get('x,y,z', rowID=idx))
 
         for idir, plane in zip([2, 1, 0], ['xy', 'xz', 'yz']):
-            sql = align_interface(self.pdb, plane=plane)
-            xyz = get_xyz_interface(sql)
+            with self.assertWarns(UserWarning) as ex:
+                db = align_interface(self.pdb, plane=plane)
+            xyz = get_xyz_interface(db)
             u, v = pca(xyz)
             vmax = np.abs(v[:, np.argmin(u)])
             assert np.argmax(vmax) == idir
 
+if __name__ == '__main__':
+    unittest.main()
