+++ pdb2sql/__init__.py
@@ -3,5 +3,7 @@ from .interface import interface
 from .StructureSimilarity import StructureSimilarity
 from . import transform
 from .utils import fetch
+from .align import align, align_interface
+from .superpose import superpose
 
 from .__version__ import __version__
