+++ CHANGELOG.rst
@@ -5,10 +5,22 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
-[Unreleased]
-************
+0.3.0
+    *****
 
 Added
 -----
 
 * Empty Python project directory structure
+
+Fixed
+=====
+
+
+
+# with overline, for parts
+* with overline, for chapters
+=, for sections
+-, for subsections
+^, for subsubsections
+", for paragraphs
