+++ CHANGELOG.md
@@ -1,0 +1,12 @@
+# Change Log
+
+
+All notable changes to this project will be documented in this file.
+This project adheres to [Semantic Versioning](http://semver.org/).
+
+## 0.3.0
+
+- Added `align` to superpose a structure to a specific axis or plane
+- Added `superpose` to superpose two structures based on selections
+- Added `fetch` to download PDB file with given PDB ID
+- Updated `interface` object to take `pdb2sql` object as input
