### 2020-03-28 5983f2b
Delete CHANGELOG.rst
