+++ pdb2sql/__init__.py
@@ -7,3 +7,8 @@ from .align import align, align_interface
 from .superpose import superpose
 
 from .__version__ import __version__
+
+# remove unnecesary modules
+del pdb2sql_base
+del pdb2sqlcore
+del utils
