### 2020-04-04 5d91af3
added align_pdb examples to exampel folder
