### 2020-04-13 #2077 @JoaoRodrigues
Hi @NicoRenaud @CunliangGeng

Thank you for addressing my points! I'd only add somewhere in the documentation page a mention to the `fetch` function. It's good for me! Approved!! Great work!
