### 2020-04-14 945e0a4
change doc to docs in setup.py
