### 2020-04-14 #2077 @CunliangGeng
> Thank you for addressing my points! I'd only add somewhere in the documentation page a mention to the `fetch` function. It's good for me! Approved!! Great work!
@JoaoRodrigues Thanks for the comment. The `fetch` has been mentioned in the docs but the docs build failed. Now it is fixed https://pdb2sql.readthedocs.io.
