### 2020-04-16 DeepRank/pdb2sql#30 @joaomcteixeira
pdb2sql read Path from pathlib
I have tried to load a pdb file from my hardisk after `fetch` and using it as a `Path` object, but `pdb2sql` cannot read a PDB from `pathlib.Path`.

**Describe the solution you'd like**

Compatibility with `pathlib` would facilitate using `pdb2sql` as a library and dependency

How to reproduce the issue:

```python
from pathlib import Path
from pdb2sql import pdb2sql, fetch

fetch('16PK')
pdbpath = Path(f'16PK.pdb')
assert pdbpath.exists()
print(pdbpath)
pdb2sql(pdbpath)

```

it breaks though the file exists in the computer:

```python
16PK.pdb
Traceback (most recent call last):
  File "/home/joao/GitHub/pdb2sql/pdb2sql/tpath.py", line 8, in <module>
    pdb2sql(pdbpath)
  File "/home/joao/anaconda3/lib/python3.7/site-packages/pdb2sql/pdb2sqlcore.py", line 26, in __init__
    self._create_sql()
  File "/home/joao/anaconda3/lib/python3.7/site-packages/pdb2sql/pdb2sqlcore.py", line 70, in _create_sql
    pdbdata = pdb2sql.read_pdb(pdbfile)
  File "/home/joao/anaconda3/lib/python3.7/site-packages/pdb2sql/pdb2sqlcore.py", line 176, in read_pdb
    raise ValueError(f'Invalid pdb input: {pdbfile}')
ValueError: Invalid pdb input: 16PK.pdb
```

Link to JOSS review: https://github.com/openjournals/joss-reviews/issues/2077
