### 2020-04-16 DeepRank/pdb2sql#31 @joaomcteixeira
Proper error message when fetching CIF files
When fetching a PDB ID that only exists in mmCIF format, the error message raised is that same when a non-existing code is given. To my experience, not accurate error messages generate frustrating experiences to a user, in this case, one who actually knows the PDB ID exists. Considering `pdb2sql` is to be used mostly as a library, I suggest adding a more specific error message, for example: `The IDB ID given is only represented in mmCIF format and pdb2sql does not handle mmCIF formats`.

**to reproduce the issue:**
```python
# https://www.rcsb.org/structure/6SL9
pdb2sql.fetch('6SL9')
```

Link to the JOSS review: https://github.com/openjournals/joss-reviews/issues/2077
