### 2020-04-16 DeepRank/pdb2sql#32 @joaomcteixeira
Allow pdb2sql to read string data
Currently, `pdb2sql` documentation state that `pdb2sql` can be initiated from `str, list or array`.

```plain
|  __init__(self, pdbfile, **kwargs)
 |      Create a SQL database with PDB data.
 |
 |      Args:
 |          pdbfile(str, list, ndarray): pdb file or data
```

However `pdb2sql` cannot read from `str` representing the PDB data.

```python
_3cro = Path('3CRO.pdb')
assert _3cro.exists()
zz = pdb2sql(_3cro.read_text())
```

when I perform this, the whole pdb data is printed to the `stdout` with a very minor message at the end:

```plain
HETATM 1884  O   HOH R  76     -25.999 -27.796  41.843  1.00 56.38           O
HETATM 1885  O   HOH R  77     -22.979 -24.940  42.923  1.00 25.75           O
MASTER      374    0    0   10    0    0    0    6 1881    4    0   16
END
 not found
```

Implementing reading from PDB string data to instantiate `pdb2sql` would be a very valuable feature because users could fetch the PDB from the database (I am not referring to the current `pdb2sql.fetch` function here) and direct the download output to `pdb2sql` without needing to save it to the disk or parsing it aroud to `list` or `array` before hand.

Link to JOSS review: https://github.com/openjournals/joss-reviews/issues/2077
