### 2020-04-16 DeepRank/pdb2sql#33 @joaomcteixeira
minor typo documentation
Very little typo on the documentation of `pdb2sql`
```
|  commit(self)
 |      Cpmmit to the database.
         *
```

Link to JOSS review: https://github.com/openjournals/joss-reviews/issues/2077
