### 2020-04-16 #2077 @joaomcteixeira
Dear @lpantano and @NicoRenaud

I have the article and the details of the project proposed. The package presented is relevant to our community and can help facilitate several implementations on a daily basis. Also, it has a good basis to allow the implementation of additional functionalities. Congratulations to the authors for this package!

In the continuation of the review process, I have raised 4 issues in the `pdb2sql` repository which refer to code implementations and which I found relevant to address before publication. Also, I list here additional minor concerns that I consider should be addressed also before acceptance.

- [ ] provide documentation at the module level: `import pdb2sql; help(pdb2sql)` . Though the documentation website states that the user should `from pdb2sql import pdb2sql`, the `pdb2sql` library level is the first place where a user at a Python prompt will look for help.  I was expecting here a general overview of how to use the library along with external references for more detailed explanations.
- [ ] bind one of the general printing actions in python to the `.pprint()` method, also provide a general `repr` message instead of the memory coordinates:
```python
In [117]: pdb_db
Out[117]: <pdb2sql.pdb2sqlcore.pdb2sql at 0x7f3164223890>

In [118]: print(pdb_db)
<pdb2sql.pdb2sqlcore.pdb2sql object at 0x7f3164223890>
```
- [ ] I found no place in the documentation that addresses the fact that `HETATM` are ignored.  It would be valuable to have explicit clarification on the decisions to which information is parsed out from the original PDB.
- [ ] I encourage the authors to write the complete kwargs options in `pdb2sql.get` function.
- [ ] documentation of `pdb2sql.exportpdb` is insuficient.
- [ ] what is the meaning of the `.commit()` method?

Have another comment for the authors, though this should be a concern that hinders publication, the authors might want to consider it for future implementations.

I found a bit anti-Pythonic that none of the methods of the `pdb2sql` core class return an instance of the same class. The return value is always a list, and this hinders functionalities like the following:

```python
pdb_db.get('*', resSeq=['1']).pprint()
```

Also, `pprint()` does not accept column `kwargs`. So, I wonder whether `pdb_db`'s `get` methods should return an instance of the same class containing just the subselection of the data.

I look forward to the reply of the authors to the issues raised.
