### 2020-04-17 3b666d2
add the case of only mmCIF format exist for a PDB ID
