### 2020-04-17 448b52c
add input support for Path instance, string and bytes of PDB data
