### 2020-04-21 9367dbd
add repr for classes
issue 36
