### 2020-04-23 c44c2c3
add arguments to print method
