### 2020-04-23 e860149
update print method in tutorial
