### 2020-04-23 dc9199c
add note about what pdb data is parsed
