### 2020-04-23 f75182a
update docstring about columns and selections
