### 2020-04-23 ec56d1d
change commit to private method
