### 2020-04-23 386a123
update doc about print
