### 2020-04-23 75b90af
fix print method with updated columns
