### 2020-04-23 d2d18c0
return class instance on call
