+++ pdb2sql/pdb2sqlcore.py
@@ -30,6 +30,11 @@ class pdb2sql(pdb2sql_base):
         if self.fix_chainID:
             self._fix_chainID()
 
+    def __call__(self, **kwargs):
+        """Returns an pdb2sql instance of the selected parts."""
+        pdb_data = self.sql2pdb(**kwargs)
+        return pdb2sql(pdb_data)
+
     def _create_sql(self):
         """Create a sql database containg a model PDB."""
         pdbfile = self.pdbfile
@@ -161,14 +166,16 @@ class pdb2sql(pdb2sql_base):
                     with open(pdbfile, 'r') as fi:
                         pdbdata = fi.readlines()
                 else:
-                    raise FileNotFoundError(f'{pdbfile} is not a file')
+                    raise FileNotFoundError(
+                        f'{pdbfile} is not a file')
             else:
                 # input is pdb content
                 if pdbfile.count('\nATOM ') > 3:
                     pdbdata = pdbfile.split('\n')
                 # invalid path
                 else:
-                    raise FileNotFoundError(f'File not found: {pdbfile}')
+                    raise FileNotFoundError(
+                        f'File not found: {pdbfile}')
         elif isinstance(pdbfile, Path):
             if not pdbfile.exists():
                 raise FileNotFoundError(f'File not found: {pdbfile}')
@@ -203,7 +210,8 @@ class pdb2sql(pdb2sql_base):
         if linelen < 80:
             pdb_line = pdb_line + ' ' * (80 - linelen)
         elif linelen > 80:
-            raise ValueError(f'pdb line is longer than 80:\n{pdb_line}')
+            raise ValueError(
+                f'pdb line is longer than 80:\n{pdb_line}')
         return pdb_line
 
     @staticmethod
@@ -445,7 +453,8 @@ class pdb2sql(pdb2sql_base):
                     # otherwise we just go on
                     else:
                         if k == 'rowID':
-                            vals = vals + tuple([int(iv + 1) for iv in v])
+                            vals = vals + \
+                                tuple([int(iv + 1) for iv in v])
                         else:
                             vals = vals + tuple(v)
                 else:
@@ -456,26 +465,33 @@ class pdb2sql(pdb2sql_base):
                         vals = vals + (v,)
 
                 # create the condition for that key
-                conditions.append(k + neg + ' in (' + ','.join('?' * nv) + ')')
+                conditions.append(
+                    k + neg + ' in (' + ','.join('?' * nv) + ')')
 
             # stitch the conditions and append to the query
             query += ' AND '.join(conditions)
 
             # error if vals is too long
             if len(vals) > self.SQLITE_LIMIT_VARIABLE_NUMBER:
-                print('\nError : SQL Queries can only handle a total of 999 values')
-                print('      : The current query has %d values' % len(vals))
+                print(
+                    '\nError : SQL Queries can only handle a total of 999 values')
+                print('      : The current query has %d values' %
+                      len(vals))
                 print('      : Hence it will fails.')
                 print(
                     '      : You are in a rare situation where MULTIPLE conditions have')
-                print('      : have a combined number of values that are too large')
+                print(
+                    '      : have a combined number of values that are too large')
                 print('      : These conditions are:')
                 ntot = 0
                 for k, v in kwargs.items():
-                    print('      : --> %10s : %d values' % (k, len(v)))
+                    print('      : --> %10s : %d values' %
+                          (k, len(v)))
                     ntot += len(v)
-                print('      : --> %10s : %d values' % ('Total', ntot))
-                print('      : Try to decrease max_sql_values in pdb2sql.py\n')
[…]
