### 2020-04-30 50b3173
updated doc and tuto for pdb2sql.__call__, align, superpose
