### 2020-04-30 #2077 @NicoRenaud
@joaomcteixeira Thanks for your review and your comments ! We have addressed your last comment about returning a class instance and implemented that in the `__call__` method of the `pdb2sql` class.
It is now possible to do the following:

```python
import pdb2sql
db = pdb2sql('1AK4.pdb')
db_chainA = db(chainID='A')
```
`db_chainA` is there a `pdb2sql` instance as you suggested. Thanks for this comment, it is very valuable to have such method.
