+++ pdb2sql/pdb2sqlcore.py
@@ -12,7 +12,7 @@ from .pdb2sql_base import pdb2sql_base
 
 class pdb2sql(pdb2sql_base):
 
-    def __init__(self, pdbfile, **kwargs):
+    def __init__(self, pdbfile, tablename='atom', **kwargs):
         """Create a SQL database with PDB data.
 
         Notes:
@@ -29,6 +29,7 @@ class pdb2sql(pdb2sql_base):
 
         # create the database
         self._create_sql()
+        self._create_table(pdbfile, tablename=tablename)
 
         # fix the chain ID
         if self.fix_chainID:
@@ -47,28 +48,32 @@ class pdb2sql(pdb2sql_base):
                     - no_chainID = ['A']
 
         Returns:
-            pdb2sql: an pb2sql instance 
+            pdb2sql: an pb2sql instance
 
         Examples:
             >>> sqldb pdb2sql('1AK4.pdb')
             >>> dbA = sqldb(chainID='A')
         """
-        pdb_data = self.sql2pdb(**kwargs)
-        return pdb2sql(pdb_data)
+
+        names = self._get_table_names()
+
+        for iN, n in enumerate(names):
+            pdb_data = self.sql2pdb(tablename=n, **kwargs)
+            if iN == 0:
+                new_db = pdb2sql(pdb_data, tablename=n)
+            else:
+                new_db._create_table(pdb_data, tablename=n)
+        return new_db
 
     def __repr__(self):
         return f'{self.__module__}.{self.__class__.__name__} object'
 
-    def _create_sql(self):
+    def _create_sql(self, tablename='ATOM'):
         """Create a sql database containg a model PDB."""
-        pdbfile = self.pdbfile
         sqlfile = self.sqlfile
 
         if self.verbose:
             print('-- Create SQLite3 database')
-
-        # size of the things
-        ncol = len(self.col)
 
         # open the data base
         # if we do not specify a db name
@@ -83,6 +88,11 @@ class pdb2sql(pdb2sql_base):
             self.conn = sqlite3.connect(sqlfile)
         self.c = self.conn.cursor()
 
+    def _create_table(self, pdbfile, tablename='ATOM'):
+
+        # size of the things
+        ncol = len(self.col)
+
         # intialize the header/placeholder
         header, qm = '', ''
         for ic, (colname, coltype) in enumerate(self.col.items()):
@@ -93,7 +103,8 @@ class pdb2sql(pdb2sql_base):
                 qm += ','
 
         # create the table
-        query = 'CREATE TABLE ATOM ({hd})'.format(hd=header)
+        query = 'CREATE TABLE {tablename} ({hd})'.format(tablename=tablename,
+                                                         hd=header)
         self.c.execute(query)
 
         # get pdb data
@@ -157,8 +168,8 @@ class pdb2sql(pdb2sql_base):
 
         # push in the database
         self.c.executemany(
-            'INSERT INTO ATOM VALUES ({qm})'.format(
-                qm=qm), data_atom)
+            'INSERT INTO {tablename} VALUES ({qm})'.format(tablename=tablename,
+                                                           qm=qm), data_atom)
 
     @staticmethod
     def read_pdb(pdbfile):
@@ -311,7 +322,7 @@ class pdb2sql(pdb2sql_base):
 
     # get the names of the columns
 
-    def get_colnames(self):
+    def get_colnames(self, tablename='atom'):
         """Get SQL column names.
 
         Returns:
@@ -321,7 +332,8 @@ class pdb2sql(pdb2sql_base):
             >>> db.get_colnames()
         """
 
-        cd = self.conn.execute('select * from atom')
+        cd = self.conn.execute(
+            'select * from {tablename}'.format(tablename=tablename))
         names = list(map(lambda x: x[0], cd.description))
         names = ['rowID'] + names
         return names
[…]
