### 2020-05-01 0a238e8
fix superpose to use intersection if needed
