### 2020-05-01 #2077 @joaomcteixeira
Dear @NicoRenaud

Thanks for all the new additions and corrections to the `pdb2sql` package. I can appreciate you addressed all of my issues, and this indeed improved the quality of your package. Also, I appreciate much that you considered the creation of new instances. Thanks!

I just found a [minor issue regarding the tests](https://github.com/DeepRank/pdb2sql/issues/47), that I would like you to address before my recommendation to acceptance.

@lpantano
The authors have addressed all my suggestions. I have raised a final issue (https://github.com/DeepRank/pdb2sql/issues/47) for the authors to address. After the authors have addressed that issue I fully recommend `pdb2sql` to publication in JOSS.

Best Wishes and Congratulations!
