### 2020-05-04 DeepRank/pdb2sql#47 @NicoRenaud
To run the test locally you can use :
```
cd test
pytest
```

you can install all the required package to run the test (i.e. coveralls, pytest and pytest-cov) with:

```
pip install -e .[test]
```

Let me know if that works for you and thanks for your input !
