### 2020-05-04 c0993ab
added test and fix issue with tablename in get_col_names
