### 2020-05-04 DeepRank/pdb2sql#47 @joaomcteixeira
I understand, so please update with `CONTRIBUTING` file with those instructions, because the current instructions have the output I referred. Sorry, I forgot to mention the `CONTRIBUTING` file in my previous message.

Regarding the mandatory requirement of running `pytest` from the `tests` folder,  you can overcome this by assigning the paths you use in tests relative to the test files themselves and not to the `cwd`.

```python
test_folder = Path(__file__).resolve().parent
```

and work from there.

Cheers and nice package! :wink:
