+++ CONTRIBUTING.rst
@@ -35,7 +35,7 @@ You want to make some kind of change to the code base
 #. (**important**) announce your plan to the rest of the community *before you start working*. This announcement should be in the form of a (new) issue;
 #. (**important**) wait until some kind of consensus is reached about your idea being a good idea;
 #. if needed, fork the repository to your own Github profile and create your own feature branch off of the latest master commit. While working on your feature branch, make sure to stay up to date with the master branch by pulling in changes, possibly from the 'upstream' repository (follow the instructions `here <https://help.github.com/articles/configuring-a-remote-for-a-fork/>`__ and `here <https://help.github.com/articles/syncing-a-fork/>`__);
-#. make sure the existing tests still work by running ``python setup.py test``;
+#. make sure the existing tests still work by running ``pytest`` from the `test/` folder;
 #. add your own tests (if necessary);
 #. update or expand the documentation;
 #. `push <http://rogerdudler.github.io/git-guide/>`_ your feature branch to (your fork of) the PDB2SQL repository on GitHub;
