### 2020-05-20 #2077 @arfon
:wave: @NicoRenaud - when do you think you might be able to address @joaomcteixeira's feedback? JOSS [is now open again](https://blog.joss.theoj.org/2020/05/reopening-joss) for submissions and we're actively trying to wrap up existing reviews.
