### 2020-05-22 #2077 @NicoRenaud
@arfon we updated the instruction to manually run the test as pointed out by @joaomcteixeira. We could also follow his suggestion to change the exec path but I don't really have the time to do that at the moment. So from our side we are done I think
