### 2020-05-22 22b4223
add paths to test folder
