+++ test/__init__.py
@@ -1,0 +1,5 @@
+"""Init module for tests package."""
+from pathlib import Path
+
+test_folder = Path(__file__).resolve().parent
+pdb_folder = Path(test_folder, 'pdb')
