+++ pdb2sql/StructureSimilarity.py
@@ -104,7 +104,8 @@ class StructureSimilarity(object):
         if lzone is None:
             resData = self.compute_lzone(save_file=False)
         elif not os.path.isfile(lzone):
-            resData = self.compute_lzone(save_file=True, filename=lzone)
+            resData = self.compute_lzone(
+                save_file=True, filename=lzone)
         else:
             resData = self.read_zone(lzone)
 
@@ -138,9 +139,13 @@ class StructureSimilarity(object):
                 self.decoy, resData, return_not_in_zone=True)
             xyz_ref_long, xyz_ref_short = self.get_xyz_zone_backbone(
                 self.ref, resData, return_not_in_zone=True)
-
+        print(resData)
+        print(data_ref_long)
+        print(data_decoy_long)
+        print(atom_long)
+        print(xyz_ref_long)
         xyz_decoy_short = superpose_selection(
-                xyz_decoy_short, xyz_decoy_long, xyz_ref_long, method)
+            xyz_decoy_short, xyz_decoy_long, xyz_ref_long, method)
 
         # compute the RMSD
         return self.get_rmsd(xyz_decoy_short, xyz_ref_short)
@@ -161,9 +166,10 @@ class StructureSimilarity(object):
         Returns:
             dict: definition of the zone.
         """
-        sql_ref = pdb2sql(self.ref)
+        sql_ref = pdb2sql(self.ref, fix_chainID=True)
         nA = len(sql_ref.get('x,y,z', chainID='A'))
         nB = len(sql_ref.get('x,y,z', chainID='B'))
+        print(nA, nB)
 
         # detect which chain is the longest
         long_chain = 'A'
@@ -264,12 +270,14 @@ class StructureSimilarity(object):
 
         # extract the xyz
         else:
-            xyz_contact_decoy = self.get_xyz_zone_backbone(self.decoy, resData)
-            xyz_contact_ref = self.get_xyz_zone_backbone(self.ref, resData)
+            xyz_contact_decoy = self.get_xyz_zone_backbone(
+                self.decoy, resData)
+            xyz_contact_ref = self.get_xyz_zone_backbone(
+                self.ref, resData)
 
         # superpose the fragments
         xyz_contact_decoy = superpose_selection(xyz_contact_decoy,
-            xyz_contact_decoy, xyz_contact_ref, method)
+                                                xyz_contact_decoy, xyz_contact_ref, method)
 
         # return the RMSD
         return self.get_rmsd(xyz_contact_decoy, xyz_contact_ref)
@@ -285,7 +293,7 @@ class StructureSimilarity(object):
         Returns:
             dict: i-zone definition
         """
-        sql_ref = interface(self.ref)
+        sql_ref = interface(self.ref, fix_chainID=True)
         contact_ref = sql_ref.get_contact_atoms(
             cutoff=cutoff, extend_to_residue=True)
 
@@ -485,8 +493,10 @@ class StructureSimilarity(object):
         backbone = ['CA', 'C', 'N', 'O']
 
         # create the sql
-        sql_decoy = pdb2sql(self.decoy, sqlfile='decoy.db')
-        sql_ref = pdb2sql(self.ref, sqlfile='ref.db')
+        sql_decoy = pdb2sql(
+            self.decoy, sqlfile='decoy.db', fix_chainID=True)
+        sql_ref = pdb2sql(
+            self.ref, sqlfile='ref.db', fix_chainID=True)
 
         # extract the pos of chains A
         xyz_decoy_A = np.array(
@@ -494,7 +504,8 @@ class StructureSimilarity(object):
                 'x,y,z',
                 chainID='A',
                 name=backbone))
-        xyz_ref_A = np.array(sql_ref.get('x,y,z', chainID='A', name=backbone))
+        xyz_ref_A = np.array(sql_ref.get(
[…]
