### 2020-05-24 44d36d8
check for chain in lzone calc
