### 2020-05-24 65b2b8f
check chainID in compute izone
