### 2020-05-24 e16bddb
raise error if more than 2 chains are found
