+++ test/test_utils.py
@@ -1,6 +1,5 @@
 import os
 import unittest
-from pathlib import Path
 from pdb2sql import fetch
 
 class TestTools(unittest.TestCase):
