### 2020-05-26 7df74a8
Update test_many2sql.py
