### 2020-05-27 #2077 @NicoRenaud
@arfon just to let you know that @joaomcteixeira has implemented his own suggestions for the test so we are officially done with the review on our side ! Thanks again @joaomcteixeira  !!
