### 2020-05-27 #2077 @lpantano
Perfect! @NicoRenaud can you confirm the latest version is 0.2.2? if that is right, can you work on a zenodo archive of this version and having the title and authors matching what is shown in the paper? Thanks!
