### 2020-05-27 #2077 @CunliangGeng
> Perfect! @NicoRenaud can you confirm the latest version is 0.2.2? if that is right, can you work on a zenodo archive of this version and having the title and authors matching what is shown in the paper? Thanks!
Thanks @lpantano, the version now is 0.4.0, and zenodo archive has been updated with your suggestions, see https://doi.org/10.5281/zenodo.3860329.
