### 2020-05-27 #2077 @danielskatz
👋 @NicoRenaud - please merge the minor changes in https://github.com/DeepRank/pdb2sql/pull/54 or let me know what you disagree with in it
